refactor: package 구조를 MVC 형태로 변경에 따른 수정
