+++ src/main/kotlin/domain/racingcar/RacingCars.kt
@@ -1,16 +1,16 @@
-package racingcar
+package domain.racingcar
 
-import number.Numbers.makeRandomsToRacing
+import domain.number.Numbers.makeRandomsToRacing
 
 class RacingCars(private var cars: Cars) {
-    fun racing(attempts: Int): Rounds {
-        val allRound = mutableListOf<Round>()
+    fun racing(attempts: Int): OverallResult {
+        val allRound = mutableListOf<Records>()
 
         repeat(attempts) {
             cars.forwardAllByCondition(makeRandomsToRacing(cars.getSize()))
             allRound.add(cars.finishRound())
         }
 
-        return Rounds(allRound)
+        return OverallResult(allRound)
     }
 }
