+++ src/main/kotlin/domain/racingcar/Record.kt
@@ -1,3 +1,3 @@
-package racingcar
+package domain.racingcar
 
 class Record(val name: String, val position: Position)
