+++ src/test/kotlin/domain/calculator/OperatorTest.kt
@@ -1,4 +1,4 @@
-package calculator
+package domain.calculator
 
 import org.assertj.core.api.Assertions.assertThat
 import org.assertj.core.api.Assertions.assertThatExceptionOfType
