+++ src/test/kotlin/domain/racingcar/PositionsTest.kt
@@ -1,4 +1,4 @@
-package racingcar
+package domain.racingcar
 
 import org.assertj.core.api.Assertions.assertThat
 import org.junit.jupiter.api.Test
